Load strings from STRINGS.CKx file
In order to better support mods/translations/etc down the line, this
patch makes many (but not all strings) dynamic: loaded as variables from
the STRINGS.CKx file on startup.

The syntax of the file is:
%string [string-name] "[string-contents]"

The string names are handcoded for each episode: some are common, others
are per-episode. They all should start "ck_str_" or "ck[456]_str_", to
distinguish episode-specific strings from strings shared between
episodes.

The string contents can contain escaped characters (currently just \n
for newlines and \\ and \" to escape the backslash and quotes), and may
be split up into multiple quote-enclosed strings (which are
concatenated), so that large strings may be split across several lines.

Note that not all strings are dynamic yet: those which are stored
directly in structures in the code (most notably the menu and the
highscores) are not converted, and there are a few others here and
there. This is largely because (a) we have to start somewhere, and (b)
the file isn't loaded when such static structures are initialised. The
current plan is to store the string names in these structures, and load
them on use.

If a string cannot be found, it will probably be replaced by its name.
This will crash if the string name is longer than the string, and the
string is US_CPrint()ed into a small area. This basically means the
[Enter] status screen crashes without the STRINGS file. This, hopefully,
will end up fixed.

Finally, there are a few other minor fixes and hacks:
- A small bugfix for the Star Wars scroller when the last line doesn't
  have a newline.
- Increases to the maximum token size.
- Got rid of some arena resets in the parser which conflicted with
  PeekToken(). Clearly we need a better way to handle that.